+++ NEWS.md
@@ -4,7 +4,7 @@
 * New function `round_coordinates()` allows to fuzzy/generalize location information by rounding deployment `latitude` and `longitude`. It also updates `coordinateUncertainty` in the deployments and `coordinatePrecision` and spatial scope in the metadata (#85).
 * `filter_deployments()` now updates the spatial and temporal scope in the metadata (#72).
 * `filter_observations()` and `filter_media()` now update the taxonomic scope in the metadata (#73).
-* `read_camtrap_dp()` now upconverts datasets to Camtrap DP 1.0.1 (#113).
+* `read_camtrap_dp()` now upgrades datasets to Camtrap DP 1.0.1. The internal function `convert()` has been renamed to `upgrade()` (#113).
 
 # camtrapdp 0.3.1